Add sorting by RAM
