+++ packages/manager/src/features/Longview/LongviewLanding/Gauges/RAM.test.tsx
@@ -2,7 +2,7 @@ import { cleanup, waitForElement } from '@testing-library/react';
 import * as React from 'react';
 import { memory } from 'src/__data__/longview';
 import { renderWithTheme } from 'src/utilities/testHelpers';
-import RAM, { generateTotalMemory, generateUsedMemory } from './RAM';
+import RAM from './RAM';
 
 afterEach(cleanup);
 
@@ -40,20 +40,6 @@ const errorStore = {
     }
   }
 };
-
-describe('Utility Functions', () => {
-  it('should generate used memory correctly', () => {
-    expect(generateUsedMemory(400, 100, 100)).toBe(200);
-    expect(generateUsedMemory(0, 100, 100)).toBe(0);
-    expect(generateUsedMemory(1000, 100, 100)).toBe(800);
-  });
-
-  it('should generate total memory correctly', () => {
-    expect(generateTotalMemory(100, 400)).toBe(500);
-    expect(generateTotalMemory(500, 400)).toBe(900);
-    expect(generateTotalMemory(100, 900)).toBe(1000);
-  });
-});
 
 describe('Longview RAM Gauge UI', () => {
   it('should render a loading state initially', () => {
