+++ packages/manager/src/features/Longview/LongviewLanding/Gauges/RAM.tsx
@@ -3,6 +3,10 @@ import * as React from 'react';
 import { WithTheme, withTheme } from 'src/components/core/styles';
 import Typography from 'src/components/core/Typography';
 import GaugePercent from 'src/components/GaugePercent';
+import {
+  generateTotalMemory,
+  generateUsedMemory
+} from '../../shared/utilities';
 import { baseGaugeProps, BaseProps as Props } from './common';
 
 import { readableBytes } from 'src/utilities/unitConversions';
@@ -114,20 +118,6 @@ const RAMGauge: React.FC<commbinedProps> = props => {
     />
   );
 };
-
-export const generateUsedMemory = (
-  used: number,
-  buffers: number,
-  cache: number
-) => {
-  /**
-   * calculation comes from original implementation of Longview.JS
-   */
-  const result = used - (buffers + cache);
-  return result < 0 ? 0 : result;
-};
-
-export const generateTotalMemory = (used: number, free: number) => used + free;
 
 export default withClientData<Props>(ownProps => ownProps.clientID)(
   withTheme(RAMGauge)
