+++ packages/manager/src/features/Longview/shared/utilities.test.ts
@@ -1,0 +1,15 @@
+import { generateTotalMemory, generateUsedMemory } from './utilities';
+
+describe('Utility Functions', () => {
+  it('should generate used memory correctly', () => {
+    expect(generateUsedMemory(400, 100, 100)).toBe(200);
+    expect(generateUsedMemory(0, 100, 100)).toBe(0);
+    expect(generateUsedMemory(1000, 100, 100)).toBe(800);
+  });
+
+  it('should generate total memory correctly', () => {
+    expect(generateTotalMemory(100, 400)).toBe(500);
+    expect(generateTotalMemory(500, 400)).toBe(900);
+    expect(generateTotalMemory(100, 900)).toBe(1000);
+  });
+});
