+++ packages/manager/src/features/Longview/shared/utilities.ts
@@ -1,4 +1,5 @@
 import { pathOr } from 'ramda';
+import { LVClientData } from 'src/containers/longview.stats.container';
 import { pluralize } from 'src/utilities/pluralize';
 import { readableBytes } from 'src/utilities/unitConversions';
 import { Disk, LongviewPackage } from '../request.types';
@@ -40,4 +41,29 @@ export const sumStorage = (DiskData: Record<string, Disk> = {}): Storage => {
     total += pathOr(0, ['fs', 'total', 0, 'y'], disk);
   });
   return { free, total };
+};
+
+export const generateUsedMemory = (
+  used: number,
+  buffers: number,
+  cache: number
+) => {
+  /**
+   * calculation comes from original implementation of Longview.JS
+   */
+  const result = used - (buffers + cache);
+  return result < 0 ? 0 : result;
+};
+
+export const generateTotalMemory = (used: number, free: number) => used + free;
+
+/**
+ * Used for calculating comparison values for sorting by RAM
+ */
+export const sumUsedMemory = (data: LVClientData) => {
+  const usedMemory = pathOr(0, ['Memory', 'real', 'used', 0, 'y'], data);
+  const buffers = pathOr(0, ['Memory', 'real', 'buffers', 0, 'y'], data);
+  const cache = pathOr(0, ['Memory', 'real', 'cache', 0, 'y'], data);
+
+  return generateUsedMemory(usedMemory, buffers, cache);
 };
